bug fixing avx

+++ include/ccmath/internal/runtime/simd/functions/impl/avx/sqrt.hpp
@@ -24,7 +24,7 @@ namespace ccm::simd
 	CCM_ALWAYS_INLINE inline simd<double, abi::avx> sqrt(simd<double, abi::avx> const & a)
 	{
 		// NOLINTNEXTLINE(modernize-return-braced-init-list)
-		return simd<float, abi::avx>(_mm256_sqrt_pd(a.get()));
+		return simd<double, abi::avx>(_mm256_sqrt_pd(a.get()));
 	}
 
 } // namespace ccm::simd
